Implement web scraping module.
Use selenium and a web driver to open an interactable web endpoint.
Query that endpoint using dynamically determined identifiers from
a set county list.

Click the download from excel button on the endpoint and rename the file
with a more specific file name.

Change .gitignore to not push the data downloaded